services/ProductTracker/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
Flask==1.1.2
requests==2.24.0
slackclient==2.9.3
slackeventsapi==2.2.1

